refactor - Moved user and auth services into a nested folder for better separation.
