+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/service/auth/MyUserDetails.java
@@ -1,4 +1,4 @@
-package dev.anirban.harmoniq_backend.service;
+package dev.anirban.harmoniq_backend.service.auth;
 
 import dev.anirban.harmoniq_backend.exception.UserNotFound;
 import dev.anirban.harmoniq_backend.repo.UserRepository;
